+++ Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/ridhoq/af"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0"
 clap = "3.0.0-beta.2"
 reqwest = { version = "0.11", features = ["json"] }
 tokio = { version = "1", features = ["full"] }
